Cleaning up.
git-svn-id: svn+ssh://cv077/home/ueshiba/svnroot/trunk/src@1602 3ef1b8b6-fb57-41bf-88d9-9a5bd220bf38

+++ TUTools++/Testprog/SURF/MatchImage.h
@@ -21,7 +21,7 @@ class MatchImage : public Image<RGB>
     Point2i	initializeH(IMG begin, IMG end)				;
     template <class IMG>
     Point2i	initializeV(IMG begin, IMG end)				;
-    MatchImage&	copy(const Image<u_char>& image, int u0, int v0)	;
+    MatchImage&	copy(const Image<u_char>& image, size_t u0, size_t v0)	;
     template <class MATCH>
     MatchImage&	drawMatches(MATCH begin, MATCH end,
 			    const Point2i& origin0,
